requirements.txt
Flask==1.1.2
pytesseract==0.3.6
mysql-connector-python==8.0.25
Pillow==7.2.0